pyproject.toml
[build-system]
requires = ["setuptools>=76.1.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"


[project]
name = "noether"
version = "0.0.0"  # only a placeholder, actual versioning happens on github
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "einops>=0.8.1",
    "wandb>=0.19.6",
    "neptune-scale>=0.10.2",
    "hydra-core>=1.3.2",
    "numpy==2.3.1",
    "torch==2.4.1",
    "torch-geometric==2.4.0",
    "torch-cluster==1.6.3",
    "torch-scatter==2.1.2",
    "pandas>=2.3.3",
    # ---
    "aistore>=1.13.5",
    "pyvista>=0.44.2",
    "rtree>=1.4.0",
    "trimesh>=4.6.6",
    # ---
    "boto3>=1.40.4",
    "huggingface-hub>=0.34.3",
    "typer>=0.16.0",
    "fsspec>=2025.5.1",
    # ---
]

[tool.setuptools_scm]
write_to = "src/noether/_version.py"

[project.scripts]
noether-train = "noether.training.cli.main_train:main"
noether-eval = "noether.inference.cli.main_inference:main"
noether-data = "noether.io.cli.cli:app"

# --- Centralized Development & Tooling Dependencies ---
# These are dependencies for developing the *entire* workspace.
# Install them with `uv sync --group dev --group docs` or `uv sync --all-groups`.
[dependency-groups]
dev = [
    {include-group = "docs"},
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov", # Added for code coverage
    "mypy[reports]>=1.15.0",
    "ruff>=0.11.0",
    "fixit>=2.1.0",
    "ipykernel>=6.30.0", # Useful for notebooks
    "types-docutils==0.22.0.20250822",
    "types-pyyaml>=6.0.12.20241230",
    "types-tqdm>=4.67.0.20250809",
    "docutils-stubs==0.0.22",
]
docs = [
    "sphinx>=7.3",
    "furo==2025.7.19",
    "sphinx-autoapi>=3.6.0",
    "sphinx-autodoc-typehints",
    "sphinx-multiversion",
    "sphinx-design",
    "myst-parser",
    "autodoc-pydantic>=2.2.0",
]

# --- Centralized UV Configuration ---
# This configuration is inherited by all sub-packages, avoiding repetition.
[tool.uv]
find-links = [
    "https://data.pyg.org/whl/torch-2.4.1+cpu.html",
    "https://data.pyg.org/whl/torch-2.4.1+cu124.html"
]
no-build-isolation-package = ["torch-scatter", "torch-cluster"]

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "torch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
    { index = "torch-cu124", marker = "sys_platform != 'darwin'" },
    { index = "torch-cpu", marker = "sys_platform == 'darwin'" },
]

# mypy and ruff ran in the child packages will fall back to this configuration
[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "torch_geometric.*",
    "torch_scatter.*",
    "torch_cluster.*",
    "aistore.*",
    "boto3.*"
#    "pyvista.*",
#    "trimesh.*",
#    "rtree.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
pythonpath = ["src"]
addopts = ["--import-mode=importlib"]

check_untyped_defs = true
explicit_package_bases = true
ignore_missing_imports = true
namespace_packages = true
no_implicit_optional = true
strict = false
warn_return_any = true
warn_unused_ignores = false
show_error_codes = true
exclude = [
    "^_build/|^build/|^dist/|^docs/|^\\.venv/",
    # avoid duplicate analysis of tests under src:
    "^src/.*/tests_.*",
    "(^|/)src/.*/(tests|tests_.*|_tests)(/|$)",
]

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "archive",
]
line-length = 120

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.ruff.lint]
# NOTE: Synchronize the ignores with .flake8
ignore = [
    "G004", # "Logging statement uses f-string" is not an issue for the current ML pipeline
    # these ignores are from flake8-bugbear; please fix!
    "B007", "B008", "B017",
    "B018", # Useless expression
    "B023",
    "B028", # No explicit `stacklevel` keyword argument found
    "E402",
    "C408", # C408 ignored because we like the dict keyword argument syntax
    "C409", # prefer [] over list()
    "C410", # prefer () over tuple()
    "E501", # E501 is not flexible enough, we're using B950 instead
    "E721",
    "E731", # Assign lambda expression
    "E741",
    "EXE001",
    "F405",
    "F841",
    # these ignores are from flake8-logging-format; please fix!
    "G101",
    # these ignores are from ruff NPY; please fix!
    "NPY002",
    # these ignores are from ruff PERF; please fix!
    "PERF203",
    "PERF401",
    "PERF403",
    # these ignores are from PYI; please fix!
    "PYI024",
    "PYI036",
    "PYI041",
    "PYI056",
    "SIM102", "SIM103", "SIM112", # flake8-simplify code styles
    "SIM113", # please fix
    "SIM105", # these ignores are from flake8-simplify. please fix or ignore with commented reason
    "SIM108", # SIM108 ignored because we prefer if-else-block instead of ternary expression
    "SIM110",
    "SIM114", # Combine `if` branches using logical `or` operator
    "SIM115",
    "SIM116", # Disable Use a dictionary instead of consecutive `if` statements
    "SIM117",
    "SIM118",
    "UP006", # keep-runtime-typing
    "UP007", # keep-runtime-typing
    "TC001",
    "TC002",
    "TC003",
]
select = [
    "B",
    "B904", # Re-raised error without specifying the cause via the from keyword
    "C4",
    "G",
    "E",
    "EXE",
    "F",
    "SIM1",
    "SIM911",
    "W",
    # Not included in flake8
    "FURB",
    "LOG",
    "NPY",
    "PERF",
    "PGH004",
    "PIE790",
    "PIE794",
    "PIE800",
    "PIE804",
    "PIE807",
    "PIE810",
    "PLC0131", # type bivariance
    "PLC0132", # type param mismatch
    "PLC0205", # string as __slots__
    "PLC3002", # unnecessary-direct-lambda-call
    "PLE",
    "PLR0133", # constant comparison
    "PLR0206", # property with params
    "PLR1722", # use sys exit
    "PLR1736", # unnecessary list index
    "PLW0129", # assert on string literal
    "PLW0133", # useless exception statement
    "PLW0406", # import self
    "PLW0711", # binary op exception
    "PLW1509", # preexec_fn not safe with threads
    "PLW2101", # useless lock statement
    "PLW3301", # nested min max
    "PT006",
    "PT022",
    "PT023",
    "PT024",
    "PT025",
    "PT026",
    "PYI",
    "Q003", # avoidable escaped quote
    "Q004", # unnecessary escaped quote
    "RSE",
    "RUF008", # mutable dataclass default
    "RUF015", # access first ele in constant time
    "RUF016", # type error non-integer index
    "RUF017",
    "RUF018", # no assignment in assert
    "RUF019", # unnecessary-key-check
    "RUF024", # from keys mutable
    "RUF026", # default factory kwarg
    "TCH",
    "TRY002", # ban vanilla raise (todo fix NOQAs)
    "TRY203",
    "TRY401", # verbose-log-message
    "UP",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.fixit]
enable = [".ci.rules"]
disable = ["fixit.rules:RewriteToLiteral"]
